move all curls to using authx methods

+++ entrypoint.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-set -Euo pipefail
+set -Euox pipefail
 
 OPA_ROOT_TOKEN=$(cat /run/secrets/opa-root-token)
 
@@ -15,36 +15,17 @@ if [[ -f "/app/initial_setup" ]]; then
     echo "initializing idp"
     python3 /app/permissions_engine/initialize_idp.py
 
-    # install base jsons at vault
-    token=$(bash get_vault_store_token.sh)
-    echo "storing data.json"
-    curl -X "PUT" "${VAULT_URL}/v1/opa/data" \
-     -H "X-Vault-Token: ${token}" \
-     -H 'Content-Type: application/json; charset=utf-8' \
-     -d @/app/data.json
+    python3 /app/initialize_vault_store.py
 
-    echo "storing access.json"
-    curl -X "PUT" "${VAULT_URL}/v1/opa/access" \
-     -H "X-Vault-Token: ${token}" \
-     -H 'Content-Type: application/json; charset=utf-8' \
-     -d @/app/permissions_engine/access.json
-
-    echo "storing paths.json"
-    curl -X "PUT" "${VAULT_URL}/v1/opa/paths" \
-     -H "X-Vault-Token: ${token}" \
-     -H 'Content-Type: application/json; charset=utf-8' \
-     -d @/app/permissions_engine/paths.json
-
-    rm /app/initial_setup
+    if [[ $? -eq 0 ]]; then
+        rm /app/initial_setup
+    fi
 fi
 
-token=$(bash get_vault_store_token.sh)
-echo $(curl "${VAULT_URL}/v1/opa/data" -H "X-Vault-Token: ${token}")
-echo $(curl "${VAULT_URL}/v1/opa/access" -H "X-Vault-Token: ${token}")
-echo $(curl "${VAULT_URL}/v1/opa/paths" -H "X-Vault-Token: ${token}")
 
 while [ 0 -eq 0 ]
 do
-  token=$(bash get_vault_store_token.sh)
+  echo "storing vault token"
+  bash get_vault_store_token.sh
   sleep 300
 done
